requirements.txt
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
beanie==1.23.6
certifi==2023.11.17
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.105.0
greenlet==3.0.2
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-model==0.2.0
Mako==1.3.0
MarkupSafe==2.1.3
motor==3.3.2
orjson==3.9.10
pydantic==2.5.2
pydantic-extra-types==2.2.0
pydantic-settings==2.1.0
pydantic_core==2.14.5
pymongo==4.6.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.27.0
toml==0.10.2
typing_extensions==4.9.0
ujson==5.9.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
